chore added a copying flag to import.sh
